requirements.txt
langchain
openai
tiktoken
pinecone-client
python-dotenv